Rename BufferUpdateManager
