+++ browser/src/neovim/NeovimBufferUpdateManager.ts
